requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.2.0
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
toml==0.10.2
